containers: support CoreOS host, for Google Compute Engine

+++ install_scripts/bootstrap_arch_linux.sh
@@ -8,6 +8,11 @@ SADM_LOCALE='en_US.UTF-8'
 SADM_CHARSET='UTF-8'
 SADM_TIMEZONE='Europe/Paris'
 
+# Arch Linux install
+ARCH_MIRROR=http://mirror.rackspace.com/archlinux
+# Release used for bootstraping from a non-Arch Linux system
+ARCH_RELEASE_DATE=2017.04.01
+
 # Usage
 if [ $# -ne 3 ]; then
     echo >&2 "Usage: $0 ROOT_DIR HOSTNAME PLAINTEXT_ROOT_PASSWORD_FILE"
@@ -21,7 +26,7 @@ if [ $# -ne 3 ]; then
 fi
 
 # Requirements checks
-if ! which arch-chroot >/dev/null; then
+if test -e /etc/arch-release && ! which arch-chroot >/dev/null; then
     echo >&2 "Error: This script requires arch-install-scripts, please run:"
     echo >&2 "> pacman -S arch-install-scripts"
     exit 1
@@ -48,10 +53,28 @@ fi
 
 # The actual Arch Linux setup starts here
 echo "[+] Installing base Arch Linux"
-pacstrap -c -d "$root_dir" base vim openssh
+if test -e /etc/arch-release; then
+  pacstrap -c -d "$root_dir" base
+else
+  (
+    cd /tmp
+    wget --continue $ARCH_MIRROR/iso/$ARCH_RELEASE_DATE/archlinux-bootstrap-$ARCH_RELEASE_DATE-x86_64.tar.gz
+    tar --strip-components=1 --directory="$root_dir" -xf archlinux-bootstrap-$ARCH_RELEASE_DATE-x86_64.tar.gz
+  )
+
+  systemd-nspawn --quiet --directory "$root_dir" /usr/bin/pacman-key --init
+  systemd-nspawn --quiet --directory "$root_dir" /usr/bin/pacman-key --populate archlinux
+fi
+
+echo "[+] Configure Arch Linux repository"
+cat >"$root_dir/etc/pacman.d/mirrorlist" <<EOF
+Server = $ARCH_MIRROR/\$repo/os/\$arch
+EOF
+
+systemd-nspawn -D "$root_dir" /usr/bin/pacman -Syu --needed --noconfirm base vim openssh rxvt-unicode-terminfo
 
 echo "[+] Configuring base system"
-echo [+] "Setting timezone to $SADM_TIMEZONE"
+echo "[+] Setting timezone to $SADM_TIMEZONE"
 ln -sf "/usr/share/zoneinfo/$SADM_TIMEZONE" "$root_dir/etc/localtime"
 
 if [[ -n $hostname ]]; then
@@ -65,14 +88,14 @@ echo "[+] Configuring locale to $SADM_LOCALE $SADM_CHARSET"
 echo "LANG=$SADM_LOCALE" > "$root_dir/etc/locale.conf"
 echo "$SADM_LOCALE $SADM_CHARSET" >> "$root_dir/etc/locale.gen"
 # There is not `locale-gen --root`, we have to use a chroot
-arch-chroot "$root_dir" locale-gen
+systemd-nspawn --quiet --directory "$root_dir" /usr/bin/locale-gen
 
 echo "[+] Setting root password"
 root_password=$(cat $root_password_file)
 if [[ -n $root_password ]]; then
   echo "root:$root_password" | chpasswd --root "$root_dir"
 else
-  echo "[+] Root password file empty, not setting root password"
+  echo "[+] Warning: root password file empty, not setting any root password"
 fi
 
 echo "[+] Disabling pam_securetty, see https://github.com/systemd/systemd/issues/852#issuecomment-127759667"
